// packages/typescript/src/generated/src/apis/CertificateOperationsApi.ts
export interface GetCertificateOperationRequest {
    operationId: string;
}

/**
 * CertificateOperationsApi - interface
 * 
 * @export
 * @interface CertificateOperationsApiInterface
 */
export interface CertificateOperationsApiInterface {
    /**
     * Creates request options for getCertificateOperation without sending the request
     * @param {string} operationId 十进制字符串异步操作 ID。
     * @throws {RequiredError}
     * @memberof CertificateOperationsApiInterface
     */
    getCertificateOperationRequestOpts(requestParameters: GetCertificateOperationRequest): Promise<runtime.RequestOpts>;

    /**
     * 返回从请求节点开始的自动重试链最新节点。
     * @summary 查询证书异步操作
     * @param {string} operationId 十进制字符串异步操作 ID。
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CertificateOperationsApiInterface
     */
    getCertificateOperationRaw(requestParameters: GetCertificateOperationRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<CertificateOperationSuccessResponse>>;

    /**
     * 返回从请求节点开始的自动重试链最新节点。
     * 查询证书异步操作
     */
    getCertificateOperation(requestParameters: GetCertificateOperationRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<CertificateOperationSuccessResponse>;

}

/**
 * 
 */
export class CertificateOperationsApi extends runtime.BaseAPI implements CertificateOperationsApiInterface {

    /**
     * Creates request options for getCertificateOperation without sending the request
     */
    async getCertificateOperationRequestOpts(requestParameters: GetCertificateOperationRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['operationId'] == null) {
            throw new runtime.RequiredError(
                'operationId',
                'Required parameter "operationId" was null or undefined when calling getCertificateOperation().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.accessToken) {
            const token = this.configuration.accessToken;
            const tokenString = await token("bearerAuth", []);

            if (tokenString) {
                headerParameters["Authorization"] = `Bearer ${tokenString}`;
            }
        }

        let urlPath = `/cert-operations/{operationId}`;
        urlPath = urlPath.replace('{operationId}', encodeURIComponent(String(requestParameters['operationId'])));

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     * 返回从请求节点开始的自动重试链最新节点。
     * 查询证书异步操作
     */
    async getCertificateOperationRaw(requestParameters: GetCertificateOperationRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<CertificateOperationSuccessResponse>> {
        const requestOptions = await this.getCertificateOperationRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => CertificateOperationSuccessResponseFromJSON(jsonValue));
    }

    /**
     * 返回从请求节点开始的自动重试链最新节点。
     * 查询证书异步操作
     */
    async getCertificateOperation(requestParameters: GetCertificateOperationRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<CertificateOperationSuccessResponse> {
        const response = await this.getCertificateOperationRaw(requestParameters, initOverrides);
        return await response.value();
    }

}