// packages/typescript/src/generated/src/apis/CertificatesApi.ts
export interface ListCertificatesRequest {
    page?: number;
    size?: number;
    keyword?: string;
    status?: CertificateStatus;
}
// ...
export enum DownloadCertificateFormatEnum {
    tar = 'tar',
    json = 'json'
}