// packages/typescript/src/generated/src/models/CertificateDetail.ts
export interface CertificateDetail {
    /**
     * 证书外部唯一标识。
     * @type {string}
     * @memberof CertificateDetail
     */
    certKey: string;
    /**
     * 
     * @type {CertificateStatus}
     * @memberof CertificateDetail
     */
    status: CertificateStatus;
    /**
     * 证书域名列表。
     * @type {Array<string>}
     * @memberof CertificateDetail
     */
    domains: Array<string>;
    /**
     * 
     * @type {CertificateAuthority}
     * @memberof CertificateDetail
     */
    ca: CertificateAuthority;
    /**
     * 
     * @type {CertificateKeyAlgorithm}
     * @memberof CertificateDetail
     */
    keyAlgorithm: CertificateKeyAlgorithm;
    /**
     * 证书生效时间。
     * @type {Date}
     * @memberof CertificateDetail
     */
    notBefore?: Date;
    /**
     * 证书过期时间。
     * @type {Date}
     * @memberof CertificateDetail
     */
    notAfter?: Date;
    /**
     * 剩余有效天数。
     * @type {number}
     * @memberof CertificateDetail
     */
    daysLeft?: number;
    /**
     * 最近失败原因。
     * @type {string}
     * @memberof CertificateDetail
     */
    failureReason?: string;
    /**
     * 记录创建时间。
     * @type {Date}
     * @memberof CertificateDetail
     */
    createdAt: Date;
    /**
     * 记录更新时间。
     * @type {Date}
     * @memberof CertificateDetail
     */
    updatedAt: Date;
}